Replace frontend with frontendV2
The original frontend did not have a environment configured to match
that of @seanflyyy's development enivronment (the issue was never
figured out) so frontendV2 was created
in order and the original files were brought over along with the data
being restructured. As a result, frontendV2 now has a better structure.

In the interest of development, frontend repository has been removed and
replaced by frontendV2.