Update TopBar file to use arrow-back-outline
This was because the original implementation were all using an icon
called chevron-left from the feather pack and they were all being
explicitly declared. So when changing from chevron-left to
arrow-back-outline, there was a need to replace all instances. Since
there are no other cases where TopBar takes in an icon other than
arrow-back-outline, it was replaced in this way.